backend/ecommerce_project/paypal_payment_app/serializers.py
from rest_framework import serializers
from .models import BillingInfo, Payment


class BillingInfoSerializer(serializers.ModelSerializer):
    class Meta:
        model = BillingInfo
        fields = '__all__'


class PaymentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Payment
        fields = '__all__'